+++ lib/features/credential/data/models/credential.dart
@@ -1,7 +1,7 @@
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:credentials/features/credential/domain/entities/credential.dart';
 
-import '../../../../src/utils/encrypter.dart';
+import '../../../../core/shared/encrypter.dart';
 
 class CredentialModel extends CredentialEntity {
   CredentialModel({
